Adding scribe usage
